Add endpoints

+++ src/Ogmas/Services/PlayersService.cs
@@ -1,3 +1,4 @@
+using System;
 using System.Collections.Generic;
 using System.Linq;
 using System.Threading.Tasks;
@@ -14,12 +15,28 @@ namespace Ogmas.Services
         private readonly IMapper mapper;
         private readonly GameParticipantsRepository gameParticipantsRepository;
         private readonly OrganizedGamesRepository organizedGamesRepository;
+        private readonly SubmitedAnswersRepository submitedAnswersRepository;
+        private readonly GamesRepository gamesRepository;
+        private readonly TaskAnswersRepository taskAnswersRepository;
 
-        public PlayersService(IMapper _mapper, GameParticipantsRepository _gameParticipantsRepository, OrganizedGamesRepository _organizedGamesRepository)
+        public PlayersService(IMapper _mapper, GameParticipantsRepository _gameParticipantsRepository, OrganizedGamesRepository _organizedGamesRepository,
+                              SubmitedAnswersRepository _submitedAnswersRepository, GamesRepository _gamesRepository, TaskAnswersRepository _taskAnswersRepository)
         {
             mapper = _mapper;
             gameParticipantsRepository = _gameParticipantsRepository;
             organizedGamesRepository = _organizedGamesRepository;
+            submitedAnswersRepository = _submitedAnswersRepository;
+            gamesRepository = _gamesRepository;
+            taskAnswersRepository = _taskAnswersRepository;
+        }
+
+        public IEnumerable<SubmitedAnswerResponse> GetPlayerAnswers(string gameId, string userId)
+        {
+            var player = gameParticipantsRepository.Filter(x => x.GameId == gameId && x.PlayerId == userId).FirstOrDefault();
+            if(player is null)
+                throw new ArgumentException("User does not participate in the game");
+            var answers = submitedAnswersRepository.GetAnswersByPlayer(player.Id);
+            return answers.Select(x => mapper.Map<SubmitedAnswerResponse>(x));
         }
 
         public IEnumerable<PlayerResponse> GetPlayers(string gameId)
@@ -31,6 +48,14 @@ namespace Ogmas.Services
         public async Task<PlayerResponse> JoinGame(string gameId, string userId)
         {
             var game = await organizedGamesRepository.Get(gameId);
+            if(game is null)
+                throw new ArgumentException("game does not exist");
+            if(game.StartTime.CompareTo(DateTime.UtcNow) <= 0)
+                throw new InvalidOperationException("game has already started");
+
+            var playerFound = gameParticipantsRepository.Filter(x => x.GameId == gameId && x.PlayerId == userId).FirstOrDefault();
+            if(!(playerFound is null))
+                throw new InvalidOperationException("player is already in game");
             var players = gameParticipantsRepository.GetParticipantsByGame(gameId).Count();
             
             var participant = new GameParticipant
@@ -47,8 +72,42 @@ namespace Ogmas.Services
 
         public async Task<PlayerResponse> LeaveGame(string playerId)
         {
+            var found = await gameParticipantsRepository.Get(playerId);
+            if(found is null)
+                throw new ArgumentException("player does not exist");
+            
             var player = await gameParticipantsRepository.Delete(playerId);
             return mapper.Map<PlayerResponse>(player);
+        }
+
+        public async Task<SubmitedAnswerResponse> SubmitAnswer(string gameId, string playerId, string questionId, string answerId)
+        {
+            var player = gameParticipantsRepository.GetParticipantByGameAndUser(gameId, playerId);
+            if(player is null)
+                throw new Exception("Player is not in the game");
+
+            var answered = submitedAnswersRepository.Filter(x => x.GameId == gameId && x.PlayerId == player.Id && x.QuestionId == questionId);
+            if(answered.Count() != 0)
+                throw new InvalidOperationException("Question is already answered");
+
+            var organizedGame = await organizedGamesRepository.Get(gameId);
+            var game = await gamesRepository.Get(organizedGame.GameTypeId);
+            if(game is null)
+                throw new Exception("Game does not exist");
+
+            var answer = taskAnswersRepository.Filter(x => x.Id == answerId && x.GameTask.Id == questionId && x.GameTask.GameId == game.Id);
+            if(answer.Count() != 1)
+                throw new Exception("Answer for this question in this game was not found");
+
+            var submited = await submitedAnswersRepository.Add(new SubmitedAnswer()
+            {
+                GameId = gameId,
+                PickedAnswerId = answerId,
+                PlayerId = player.Id,
+                QuestionId = questionId
+            });
+
+            return mapper.Map<SubmitedAnswerResponse>(submited);
         }
     }
 }
